Correct the publishing, gradle plugins must be published slightly differently
In https://github.com/git-commit-id/git-commit-id-gradle-plugin/issues/1 I noted that
this plugin likely needs the publishing outlined in https://github.com/davidweber411/how-to-publish-to-maven-central-with-gradle/.
The main reason is that maven plugins are getting published to maven central.

However I now believe that especially gradle plugins may not to be configured published to maven.central.

After reading
https://docs.gradle.org/current/userguide/java_gradle_plugin.html#maven_publish_plugin
https://docs.gradle.org/current/userguide/publishing_gradle_plugins.html

it became clear that gradle plugins should use 'com.gradle.plugin-publish' for publications, instead of a 'maven-publish'.
Furthermore the GradlePluginDevelopmentExtension has a configuration to automatedPublishing.
When set to true gradle will create the relevant publication tasks.
With the extra publication added from https://github.com/davidweber411/how-to-publish-to-maven-central-with-gradle/
it made suddenly perfectly sense that gradle was giving me a warning messages that the mainJava publication was overwritten.

there is one publication for the main java component of the plugin,
but there are also pom publications for each of the apply '...' things that then reference the plugin.
That way a plugin can provide multiple plugins that can be applied, but internally gradle will 'just' reference the same jar.

From the gradle docs:
```
When the Java Gradle Plugin (java-gradle-plugin) detects that the Maven Publish Plugin (maven-publish) is also applied by the build,
it will automatically configure the following MavenPublications:

- a single "main" publication, named pluginMaven, based on the main Java component
- multiple "marker" publications (one for each plugin defined in the gradlePlugin {} block),
  named <pluginName>PluginMarkerMaven (for example in the above example it would be simplePluginPluginMarkerMaven)
```