Avoid injecting a project in GitCommitIdPluginGenerationTask
A build with gradle 5.4.1 is failing with
```
* What went wrong:
A problem occurred evaluating root project 'git-commit-id-gradle-debugging.NiPqb7ARWM'.
> Failed to apply plugin [id 'io.github.git-commit-id.git-commit-id-gradle-plugin']
   > Could not create task ':gitCommitIdGenerationTask'.
      > No service of type Project available in ProjectScopeServices.
```
This is likely due to the fact that the GitCommitIdPluginGenerationTask is trying to inject a project.
However I believe this is not even needed.
Why?
The expectation is that lazy properties are only supported with gradle 5.1 so we can treat it as minimal supported version.

As per https://docs.gradle.org/5.1/javadoc/org/gradle/api/DefaultTask.html
the DefaultTask inherits a Task.
And as per
https://docs.gradle.org/5.1/javadoc/org/gradle/api/Task.html#getProject()
a task already has the ability to get the project.

Hence there should be no need to inject a project into a GitCommitIdPluginGenerationTask / DefaultTask / Task.