Experiment with up-to-date checks and exports of properties to the project
This is yet another idea of experting the generated properties to the project
while keeping the task that generates them as "up-to-date".

in this implementation the task essentially generates two properties.
One is the desired one configured by the user,
the other is the "internal" one used to expose the properties to the project.

The concept relies on groovy closures.
Not sure if this ideal.